+++ src/store/tasks.ts
@@ -48,6 +48,11 @@ const tasksSlice = createSlice({
 		builder.addCase(thunks.tasks.cancel.rejected, (_, action) => {
 			log.error('Error occured while trying to cancel a task', action.error.message);
 		});
+		builder.addCase(thunks.tasks.removeTask.fulfilled, (state, action) => {
+			const copy = state.tasks;
+			delete copy[action.payload];
+			state.tasks = copy;
+		});
 		builder.addCase(thunks.posts.downloadPosts.pending, (state, action) => {
 			const newId = state.lastId + 1;
 			state.lastId = newId;
@@ -62,15 +67,21 @@ const tasksSlice = createSlice({
 		});
 		builder.addCase(thunks.posts.downloadPost.fulfilled, (state, action) => {
 			const id = action.meta.arg.taskId;
+			log.debug('UPDATING TASK');
 			if (id) {
 				const task = state.tasks[id];
 				task.itemsDone = task.itemsDone + 1;
-				if (task.itemsDone / task.items >= 1) {
-					task.state = 'completed';
-					task.timestampDone = Date.now();
-				} else if (task.state !== 'canceled') {
+
+				if (task.state !== 'canceled') {
 					task.state = 'downloading';
 				}
+			}
+		});
+		builder.addCase(thunks.posts.persistTask.fulfilled, (state, action) => {
+			const task = action.payload;
+
+			if (task) {
+				state.tasks[task.id] = task;
 			}
 		});
 	},